requirements.txt
fastapi
uvicorn
streamlit
langchain
google-api-python-client
google-auth
google-auth-oauthlib
python-dotenv
httpx
